Completed assignment

+++ FVCalc.java
@@ -3,8 +3,8 @@ public class FVCalc {
 	public static void main(String[] args){
 	int invested = Integer.parseInt(args[0]);
 	double interestrate = Double.parseDouble(args[1]);
-	double years = Double.parseDouble(args[2]);
-	int futurevalue = invested * (int)(Math.pow((1 + interestrate/100), years));
+	int years = Integer.parseInt(args[2]);
+	double futurevalue = invested * (double)(Math.pow((1 + interestrate/100), years));
 	System.out.println("The future value of an investment of " + invested + " Shekels after " + years + " years at an interest rate of " + interestrate + "% is " + futurevalue + " Shekels.");
 	}
 }
